requirements.txt
flask
numpy
scikit-learn
scipy
matplotlib
Flask==2.2.3
Werkzeug==2.2.3
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1
open_clip_torch==2.16.0
torch>=2.0.0